Implemented periodic editing of move progress indicators.

+++ Classes/GameMovementHandler.ts
@@ -95,17 +95,19 @@ export default class GameMovementHandler {
     /**
      * Caches the player set and creates a move timer for them.
      * @param players - The players to create a move timer for.
-     * @param sendProgressIndicators - Whether or not to send messages to each player with progress indicators.
      * @param callback - The callback to execute on each tick of the move timer.
+     * @param sendProgressIndicators - Whether or not to send messages to each player with progress indicators. Defaults to false.
+     * @param time - The number of milliseconds it will take to move to the destination. Required when sendProgressIndicators is true.
      */
-    #createMoveTimerFor(players: Set<Player>, sendProgressIndicators: boolean, callback: () => Promise<void>) {
+    #createMoveTimerFor(players: Set<Player>, callback: () => Promise<void>, sendProgressIndicators: boolean = false, time?: number): void {
         this.#addPlayerSet(players);
         this.#moveTimeRatios.set(players, 0);
+        this.#moveTimers.set(players, setInterval(callback, Game.tick));
         if (sendProgressIndicators) {
             const progressIndicator = this.#generateProgressIndicator(players);
             this.#game.communicationHandler.sendMoveProgressIndicatorToPlayers(players, progressIndicator);
+            this.#createMoveProgressTimerFor(players, time);
         }
-        this.#moveTimers.set(players, setInterval(callback, Game.tick));
     }
 
     /**
@@ -144,7 +146,7 @@ export default class GameMovementHandler {
             // Clean up the associated player set.
             for (const player of players) {
                 this.#indexedPlayerSets.delete(player.name);
-                this.deleteMoveProgressIndicator(player);
+                this.#deleteMoveProgressIndicator(player);
             }
             this.#playerSets.delete(players);
         }
@@ -170,14 +172,12 @@ export default class GameMovementHandler {
      * Deletes the player's move progress indicator from the cache and its associated message.
      * @param player - The player whose move progress indicator is to be deleted.
      */
-    public async deleteMoveProgressIndicator(player: Player): Promise<void> {
+    async #deleteMoveProgressIndicator(player: Player): Promise<void> {
         const moveProgressIndicator = this.#moveProgressIndicators.get(player.name);
         if (!moveProgressIndicator) return;
         this.#moveProgressIndicators.delete(player.name);
-        const channel = await this.#game.clientContext.client.channels.fetch(moveProgressIndicator.channelId);
-        if (!channel.isTextBased()) return;
-        const message = await channel.messages.fetch(moveProgressIndicator.messageId);
-        await message.delete().catch();
+        const message = await this.#game.clientContext.getSentMessage(moveProgressIndicator);
+        if (message) await message.delete().catch();
     }
 
     /**
@@ -232,6 +232,18 @@ export default class GameMovementHandler {
     }
 
     /**
+     * Determines whether to send a move progress indicator.
+     * We don't want players to be able to predict if a heated situation is occurring, so the result is randomly determined.
+     * Returns true if the amount of time it takes to complete the movement in milliseconds is greater than a
+     * randomly generated number within a given range.
+     * @param time - The number of milliseconds it will take to move to the destination.
+     */
+    private doSendProgressIndicator(time: number): boolean {
+        const timeThreshold = this.#game.rollCustomDie(8, 12);
+        return time / 1000 >= timeThreshold.result;
+    }
+
+    /**
      * Generates a string containing a progress indicator bar.
      * @param players - The set of players to generate a progress indicator for.
      * @param width - The number of characters comprising the progress indicator. Defaults to 16.
@@ -246,6 +258,46 @@ export default class GameMovementHandler {
         const filledCount = Math.floor(ratio * width);
         const bar = fillChar.repeat(filledCount) + emptyChar.repeat(width - filledCount);
         return `[${bar}]`;
+    }
+
+    /**
+     * Calculates the interval with which to edit the players' move progress indicators based on
+     * how long the movement is expected to take. The interval will be randomly generated within
+     * a given range, to make it more difficult for players to predict if a heated situation is occurring.
+     * @param time - The number of milliseconds it will take to move to the destination.
+     */
+    #calculateMoveProgressInterval(time: number): number {
+        // Edit every 5 seconds, at most.
+        const maxInterval = 5000;
+        // Decide the target number of steps. Most movements are short, so we only need a few steps.
+        // Just in case it's a longer movement, we'll add significantly more steps.
+        const longMovement = time > 60000;
+        const steps = longMovement ? 30 : 8;
+        // Ensure the interval is never lower than the maxInterval.
+        const baseInterval = Math.max(maxInterval, time / steps);
+        // Make the interval vary by up to 25% to make it less predictable.
+        const variationPercent = this.#game.rollCustomDie(75, 125).result;
+        const heatedSlowdownMultiplier = this.#game.heated ? 1 + this.#game.settings.heatedSlowdownRate : 1;
+        return Math.max(maxInterval, heatedSlowdownMultiplier * baseInterval * (variationPercent / 100));
+    }
+
+    /**
+     * Creates an interval timer that periodically edits the players' move progress indicator
+     * messages to reflect the current ratio of the movement's completion.
+     * @param players - The set of players to create a progress timer for.
+     * @param time - The total movement duration in milliseconds.
+     */
+    #createMoveProgressTimerFor(players: Set<Player>, time: number): void {
+        const interval = this.#calculateMoveProgressInterval(time);
+        this.#moveProgressTimers.set(players, setInterval(async () => {
+            const progressIndicator = this.#generateProgressIndicator(players);
+            for (const player of players) {
+                const sentMessage = this.#moveProgressIndicators.get(player.name);
+                if (!sentMessage) continue;
+                const message = await this.#game.clientContext.getSentMessage(sentMessage);
+                this.#game.communicationHandler.editMessage(message, progressIndicator);
+            }
+        }, interval));
     }
 
     /**
@@ -295,7 +347,8 @@ export default class GameMovementHandler {
         }
         const startingPos = this.#getPosition(players);
 
-        this.#createMoveTimerFor(players, false, async () => {
+        const sendProgressIndicator = this.doSendProgressIndicator(time);
+        this.#createMoveTimerFor(players, async () => {
             const settings = this.#game.settings;
             const firstPlayer = this.#first(players);
             let remainingTime = firstPlayer.remainingTime;
@@ -412,7 +465,7 @@ export default class GameMovementHandler {
                     if (time > 1000) this.#game.narrationHandler.narratePartyReady(action, firstPlayer.party.leader);
                 }
             }
-        });
+        }, sendProgressIndicator, time);
     }
 
     /**
@@ -426,7 +479,7 @@ export default class GameMovementHandler {
         if (players.size === 0) return;
         for (const player of players)
             player.remainingTime = delay;
-        this.#createMoveTimerFor(players, false, async () => {
+        this.#createMoveTimerFor(players, async () => {
             const settings = this.#game.settings;
             const firstPlayer = this.#first(players);
             let remainingTime = firstPlayer.remainingTime;
